requirements.txt
langchain 
chromadb 
python-dotenv 
sentence-transformers 
langchain-community 
langchain-huggingface